+++ src/components/FunctionsPage/Functions.js
@@ -60,7 +60,6 @@ import { useGroupContent } from '../../hooks/groupContent.hook'
 import { useMode } from '../../hooks/mode.hook'
 import { useVirtualization } from '../../hooks/useVirtualization.hook'
 import { useYaml } from '../../hooks/yaml.hook'
-import { useTableScroll } from '../../hooks/useTableScroll.hook'
 
 import cssVariables from './functions.scss'
 
@@ -697,17 +696,10 @@ const Functions = ({
       headerRowHeight: cssVariables.functionsHeaderRowHeight,
       rowHeight: cssVariables.functionsRowHeight,
       rowHeightExtended: cssVariables.functionsRowHeightExtended
-    }
+    },
+    activateTableScroll: true
   })
 
-  useTableScroll({
-    content: tableContent,
-    rowHeight: cssVariables.functionsRowHeight,
-    rowHeightExtended: cssVariables.functionsRowHeightExtended,
-    selectedItem: selectedFunction,
-    expandedRowsData: selectedRowData,
-    tableRef
-  })
 
   return (
     <FunctionsView
